+++ release_notes.md
@@ -3,3 +3,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+- Fixed an issue where custom handler apps with extension bundles would fail to load binding types after specialization (#11676)